# backend/openapi.yaml
openapi: 3.1.0
info:
  title: AI Threat Modeler API
  description: |
    API for the AI Threat Modeler Dashboard - A comprehensive application security platform 
    with threat modeling and AI-powered chat capabilities.
    
    ## Authentication
    Most endpoints require JWT Bearer token authentication. Obtain a token by logging in 
    via the `/api/auth/login` endpoint.
    
    ## User Roles
    - **Admin**: Full access to all features including user management and settings
    - **Operator**: Can schedule and manage their own jobs (threat modeling)
    - **Auditor**: Read-only access to view all jobs and reports
    
  version: 0.9.0
  contact:
    name: Sam Li
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: http://localhost:3001/api
    description: Local development API
...
security:
  - BearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Create a new user account. First user becomes Admin automatically.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                  minLength: 3
                  example: john_doe
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  minLength: 6
                  format: password
                  example: securePassword123
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User created successfully
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Username or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate user and receive JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  description: Username or email
                  example: admin
                password:
                  type: string
                  format: password
                  example: password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Login successful
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Retrieve authenticated user's information
      responses:
        '200':
          description: User information retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/auth/change-password:
    post:
      tags:
        - Authentication
      summary: Change password
      description: Change the authenticated user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - currentPassword
                - newPassword
              properties:
                currentPassword:
                  type: string
                  format: password
                newPassword:
                  type: string
                  format: password
                  minLength: 6
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password changed successfully
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
...
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: admin
        email:
          type: string
          format: email
          example: admin@example.com
        role:
          $ref: '#/components/schemas/UserRole'
        password_changed:
          type: boolean
          description: Whether user has changed their default password
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserRole:
      type: string
      enum:
        - Admin
        - Operator
        - Auditor
      description: |
        User roles:
        - Admin: Full system access including user management and settings
        - Operator: Can schedule and manage jobs
        - Auditor: Read-only access to view all jobs and reports
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Error message
        message:
          type: string
          example: Detailed error description
        code:
          type: string
          description: Optional error code
          example: USER_NOT_FOUND
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Bad request
            message: Invalid parameters provided
    Unauthorized:
      description: Unauthorized - missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Access token required
            message: Please provide a valid authentication token
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Not found
            message: The requested resource was not found
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Internal server error
            message: An unexpected error occurred